Fix spurious detection of alire.toml during build test (#201)
The detection is correct in the sense that the file is at ./alire/alire.toml,
but it is not an old manifest but the proper manifest of the alire repo. This
is an unexpected and unwanted interaction that should go away once the check
for old manifests is removed. For now, it's simpler to just do the tests in a
clean folder.