Rename GNAT crates
I changed the names of the crates to what I think is more meaningful.

I  merged the Linux, Windows and macOS under a gnat_native that will be
easier to understand and use in my opinion.

I also added a licenses field.

One thing that we may want to improve is the values for the case
statements, for instance `[origin."case(os)".macos."case(word-size)".bits-64]`
will soon be a blocking point to support macOS on the M1.

Instead of word size we probably need something like: "case(CPU)" or
"case(arch)". Or even use full platform triplets: "case(host_triplet)".