Tash and scripted_testing (#737)
* New release tash-8.7.1.

The main change in this release is replacing options.gpr, which was
not distinguished enough and could conflict with other crates. It was
never intended to be used outside this crate, and the function is
replaced by tcl_tk_options.gpr.

* New crate scripted_testing-2.0.0.

This crate supports scriptable integration testing, where the mocked
environment of the software being tested can be scripted. The
supported scripting language is Tcl.

It's intended to add some automation to the development of mocking
support, probably using libadalang to generate stubs.