Added base package for Ada ISO Library
This is my first time publishing to alire, so please let me know if I made any mistakes. This is the base (mostly empty) package but I'm hoping for it to be a dependency for two other alire items I'll be publishing (the almost-completed iso_countries https://github.com/ada-iso/iso_countries and the soon-to-be worked on iso_curriences).  Before I attempt to publish iso_countries I'd like to have iso_base already in there.  Please let me know if you have any questions!