libreadline macOS external (#976)
* Add macOS support to libreadline.

"brew info readline" shows that the library is "keg-only", i.e. it's
not symlinked into Homebrew's normally-visible include/, lib/
folders. This doesn't matter for the main usage in Alire
(gnatcoll_readline), because it's only requirement is to link against
"-lreadline". On macOS this is provided by BSD libedit.

In that case, why bother installing it? To stop alr warning
"Generating possibly incomplete environment because of missing
dependencies".

  * index/li/libreadline/libreadline-external.toml: For Homebrew & MacPorts,
      the external package is "readline".

* Add libreadline support for Centos, Fedora

  * index/li/libreadline/libreadline-external.toml: as summary.